gh-95: Update to Swagger v0.2.4

+++ models/locations_list_response.go
@@ -35,7 +35,7 @@ type LocationsListResponse struct {
 	Locations []*LocationGetResponse `json:"locations"`
 
 	// The total number of locations for the query. The number of items in a response may be smaller due to paging.
-	TotalResults int32 `json:"totalResults,omitempty"`
+	TotalResults int64 `json:"totalResults,omitempty"`
 }
 
 // Validate validates this locations list response
